+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2025-09-10
+
+### Added
+- Can use AbortSignal to terminate `parseStream` and `useJSONStream`
+- CI and CD through GitHub Actions
+
+### Features
+- `useJSONStream` React hook for easy integration with React applications.
+- New export path `/react` for the React hook.
+
 ## [3.0.0] - 2025-08-22
 
 ### Changes